JEP 455 FTW!

+++ src/test/java/com/github/beatngu13/playground/misc/PatternMatchingForSwitchTest.java
@@ -8,7 +8,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 /*
  * See:
- * - https://openjdk.org/jeps/441
+ * - https://openjdk.org/jeps/455
  * - https://twitter.com/TheDonRaab/status/1647805135869943814
  */
 class PatternMatchingForSwitchTest {
@@ -17,10 +17,10 @@ class PatternMatchingForSwitchTest {
 	void test_fizz_buzz() {
 		var list = IntStream.rangeClosed(1, 15)
 				.mapToObj(value ->
-						switch ((Integer) value) {
-							case Integer i when i % 15 == 0 -> "FizzBuzz";
-							case Integer i when i % 5 == 0 -> "Buzz";
-							case Integer i when i % 3 == 0 -> "Fizz";
+						switch (value) {
+							case int i when i % 15 == 0 -> "FizzBuzz";
+							case int i when i % 5 == 0 -> "Buzz";
+							case int i when i % 3 == 0 -> "Fizz";
 							default -> Integer.toString(value);
 						}
 				)
